version 0.3 MergeFiles.py now merges multiple files of the same barcode into one file. Merged barcode files will be saved in the same directory as the files that they were created from.
CountReads.py now creates a csv file instead of a text file. This csv will be saved in the same location as  the initial_directory selected.               ISSUES: If the file selection is cancelled, a PermissionError is thrown in the CountReads.py file. To fix this, first check that the file can be read. If it cannot be read, tell the user something went wrong and needs to be fixed. Give them another file open dialog.